add function to compute mean distance
This commit primarily adds in function to compute mean Euclidean
distance given a list of AfidDistanceSet objects and implements the
associated tests.

Additionally, makes some minor changes to the afid_sets strategy, namely
the coordinate system list it draws from. Should only be 'LPS' or 'RAS'
once it has been converted to an ``AfidSet``.

Lastly, adds in checks to make sure the inputs to the metric are of the
correct type and if necessary, whether there is a common element.