Add orthoview plots for afids
This commit adds the functionality to create a view object with AFIDs
overlaid on an nifti image. To do so, a nifti image with just the
fiducials are created with an intensity value equal to the labels. This
serves as a "segmentation" or "label" map. A custom 32-discrete colormap
object is created to visualize these different fiducials. Additionally,
to allow for either voxel coordinates or positional coordinates to be
passed, a conversion to voxel coordinates is performed when the plotting
function is called. Note, plotting is not limited to complete `AfidSet`s.
Individual and incomplete `AfidSet`s can also be plotted.

Finally stubs have been created for `matplotlib` and `nilearn` to
correct for pyright errors. `# pyright: ignore` was added where
appropriate to bypass issues false errors.