Add connectome summary plotting functionality
This commit adds the functionality to generate a summary plot of AFID
distances on a connectome using nilearn.plotting to do so. In doing so,
a complete set of AfidSet distances are provided as an argument, and
coloured based on their distance from "expected" (ground-truth).

Approporiate unit tests are also added in to make sure plots are generated
and of the correct type. During development, generated plots were visually
assessed to ensure generated plots are as expected.

Note: a `pyright: ignore` has been temporarily added to line 230 in
plotting.py until other plot_types are implemented in order to bypass
the pyright error being thrown.