Take actual value of distances (instead of abs)
Knowing if the "distance" is negative can provide some insight - this is
mostly for plotting distances along specific spatial components (e.g. x,
y, z).

These distances will be automatically set in the histogram and upcoming
scatter plots. For the connectome plot, the minimum node value is set by
either taking the smallest value (if negative) or 0 (if positive).