Add scatter summary plotting functionality
This commit adds functionality to generate a summary scatter plot of
AFID distances using `plotly`. Similar to generating a histogram plot,
this also makes use of the new `afid_labels` argument to label the
markers. If not porvided, indices are used to label the scatter plot.

Appropriate unit tests are alos added, with generated plots visually
assessed during development.